ДОбавил picturefill в проект и сборку js в автоматизацию
