+++ package.json
@@ -9,6 +9,7 @@
     "browser-sync": "^2.26.3",
     "del": "^3.0.0",
     "gulp": "4.0.x",
+    "gulp-concat": "^2.6.1",
     "gulp-csso": "^3.0.1",
     "gulp-imagemin": "^4.1.0",
     "gulp-minify": "^3.1.0",
@@ -49,7 +50,6 @@
   },
   "dependencies": {
     "node-normalize-scss": "^8.0.0",
-    "node.normalize.scss": "^3.0.3",
     "normalize-scss": "^7.0.1",
     "npx": "^10.2.0"
   }
