+++ source/js/picturefill.js
@@ -1,19 +1,3 @@
-function toggleMenu(e) {
-  var topMenu = document.querySelector('.main-nav');
-  e.preventDefault();
-
-  if (topMenu.classList.contains('main-nav--opened')) {
-    topMenu.classList.remove('main-nav--opened');
-    topMenu.classList.add('main-nav--closed');
-  } else {
-    topMenu.classList.add('main-nav--opened');
-    topMenu.classList.remove('main-nav--closed');
-  }
-}
-
-var menuButton = document.querySelector('.main-nav__toggle');
-menuButton.addEventListener('click', toggleMenu);
-
 /*! picturefill - v3.0.2 - 2016-02-12
  * https://scottjehl.github.io/picturefill/
  * Copyright (c) 2016 https://github.com/scottjehl/picturefill/blob/master/Authors.txt; Licensed MIT
